+++ DI/Details/FunctionTraits.h
@@ -21,7 +21,7 @@ namespace DI
 				operator U&() const {};
 			};
 
-			template<class T, int>
+			template<class T, size_t>
 			using WrapType = T;
 
 			template<class T, class... As>
@@ -91,10 +91,10 @@ namespace DI
 
 		}
 
-		template<class T, int Max = 10>
+		template<class T, size_t Max = 10>
 		struct FunctionArity
 		{
-			static constexpr int value = FunctionTraits::FunctionArityImpl<FunctionTraits::IsCallableCheck, T, std::make_index_sequence<Max>>::value;
+			static constexpr size_t value = FunctionTraits::FunctionArityImpl<FunctionTraits::IsCallableCheck, T, std::make_index_sequence<Max>>::value;
 		};
 
 		template<class T>
